Improve performance for inserts
The default insert implementation traverses all pages looking for
a page with enough space to add the tuple. This gives an O(N^2)
runtime for adding N tuples, which is very slow. This method also
requires O(N^2) page reads, which is also slow.

The new method uses a linked-list to keep track of the free pages,
reducing the runtime and page reads to almost linear (not counting
bookkeeping). More implementation details can be found in the
HeapTupleFile.addTuple() javadoc.

Also included is a .txt of the runtime/perf stats for a few cases.