Implement decorrelation for two cases
Decorrelates second and third cases: i.e.
  SELECT ... WHERE a IN (SELECT ...)
and
  SELECT ... WHERE EXISTS (SELECT ...)

Only implemented when inner subquery reads from BASE_TABLE.

Fixed implementation for SEMIJOIN. Now creates a new Tuple instead
of just returning the lhs Tuple. This is necessary for accurate
pinning/unpinning tracking.

Had to make expr in class InSubqueryOperator public so that it
is accessible when decorrelating.